reafctored to include create organizxer accoiunt script in seeding flow
