+++ seeding/seeds/seeds.js
@@ -1,6 +1,7 @@
 const { setupGitHubSecrets } = require("./setup-github-secrets");
+const { createOrganizerAccount } = require("./create-organizer-account");
 
-const seeds = [setupGitHubSecrets];
+const seeds = [setupGitHubSecrets, createOrganizerAccount];
 
 module.exports = {
   seeds,
